+++ packages/types/src/index.ts
@@ -39,7 +39,7 @@ export type LockConfiguration = {
 	acquireFailTimeout: number;
 	stale: number;
 	renewInterval: number;
-	runningTimeout: number;
+	maxHoldTime: number;
 	onLockLost: (lockName: string, reason: "renewFailed" | "timeout") => void;
 	onEvent: (event: LockEvent) => void;
 };